extracted LinkTravelTimeContainer to a separate class
