+++ src/main/scala/beam/router/r5/R5RoutingWorker.scala
@@ -1,14 +1,8 @@
 package beam.router.r5
 
-import java.io.InputStreamReader
-import java.io.BufferedReader
-import java.io.FileInputStream
-import java.io.InputStream
-import java.util.zip.GZIPInputStream
-
 import java.time.temporal.ChronoUnit
 import java.time.{ZoneId, ZonedDateTime}
-import java.{lang, util}
+import java.util
 
 import akka.actor._
 import akka.pattern._
@@ -26,15 +20,14 @@ import beam.router.r5.R5RoutingWorker.{R5Request, TripWithFares}
 import beam.router.r5.profile.BeamMcRaptorSuboptimalPathProfileRouter
 import beam.router.{Modes, RoutingModel}
 import beam.sim.BeamServices
-import beam.sim.metrics.{Metrics, MetricsSupport, PerformanceStats}
+import beam.sim.metrics.{Metrics, MetricsSupport}
 import com.conveyal.r5.api.ProfileResponse
 import com.conveyal.r5.api.util._
 import com.conveyal.r5.profile._
 import com.conveyal.r5.streets.{EdgeStore, StreetRouter, TravelTimeCalculator, TurnCostCalculator}
 import com.conveyal.r5.transit.{RouteInfo, TransportNetwork}
 import com.google.common.cache.{CacheBuilder, CacheLoader}
-import org.matsim.api.core.v01.network.{Link, Network}
-import org.matsim.api.core.v01.population.Person
+import org.matsim.api.core.v01.network.Network
 import org.matsim.api.core.v01.{Coord, Id}
 import org.matsim.core.router.util.TravelTime
 import org.matsim.vehicles.Vehicle
@@ -44,85 +37,6 @@ import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 import scala.concurrent.Future
 import scala.language.postfixOps
-
-
-// map is a map from slotId -> TraveTime
-case class LinkTravelTime(link: Id[Link], map: util.HashMap[Int, Double])
-
-class LinkTravelTimeContainer(fileName: String,
-                              timeBinSizeInSeconds: Int) extends TravelTime{
-
-  var linkTravelTimeMap: util.HashMap[Id[Link], LinkTravelTime] = new util.HashMap()
-
-
-
-  def loadLinkStats() = {
-    System.out.println("fileName -> " + fileName + " is being loaded")
-
-
-    val fileStream = new FileInputStream(fileName)
-    val gzipStream = new GZIPInputStream(fileStream)
-    val decoder = new InputStreamReader(gzipStream)
-    val bufferedReader = new BufferedReader(decoder)
-
-    var line: String = null
-
-
-    while ({line = bufferedReader.readLine; line != null}){
-      val linkStats = line.split(",")
-
-      if(linkStats.length == 10) {
-        val linkId = linkStats(0)
-        val hour = linkStats(3)
-        val travelTime = linkStats(9)
-        val stat = linkStats(7)
-
-        if (stat.equalsIgnoreCase("avg")) {
-          val _linkId = Id.createLinkId(linkId)
-          if (linkTravelTimeMap.containsKey(_linkId)) {
-
-            val linkTravelTime = linkTravelTimeMap.get(_linkId)
-            linkTravelTime.map.put(hour.toDouble.toInt, travelTime.toDouble)
-            linkTravelTimeMap.put(_linkId, linkTravelTime)
-          } else {
-
-            val map = new util.HashMap[Int, Double]()
-            map.put(hour.toDouble.toInt, travelTime.toDouble)
-
-            val linkTravelTime = new LinkTravelTime(_linkId, map)
-            linkTravelTimeMap.put(_linkId, linkTravelTime)
-          }
-        }
-      }
-
-    }
-
-    System.out.println("LinkTravelTimeMap is initialized")
-
-  }
-
-
-
-  def getLinkTravelTime(link: Link, time: Double, person: Person, vehicle: Vehicle): Double = {
-    if (linkTravelTimeMap.keySet().contains(link.getId)){
-      if (linkTravelTimeMap.get(link.getId).map.containsKey(getSlot(time))){
-
-        linkTravelTimeMap.get(link.getId).map.get(getSlot(time))
-      } else {
-        link.getFreespeed
-      }
-    } else {
-      link.getFreespeed
-    }
-  }
-
-  def getSlot(time: Double): Int ={
-
-    Math.round(Math.floor(time/timeBinSizeInSeconds)).toInt
-  }
-
-  loadLinkStats()
-}
 
 class R5RoutingWorker(val beamServices: BeamServices, val transportNetwork: TransportNetwork, val network: Network, val fareCalculator: FareCalculator, tollCalculator: TollCalculator) extends Actor with ActorLogging with MetricsSupport {
   val distanceThresholdToIgnoreWalking = beamServices.beamConfig.beam.agentsim.thresholdForWalkingInMeters // meters
